# api/openapi.yaml
openapi: 3.1.0
info:
  title: Simple-VPN-Builder Control Plane API
  version: 0.1.0v
  description: High-performance VPN orchestration and client protocol provisioning control plane.
servers:
  - url: http://localhost:8080
    description: Local Development Server
paths:
...
  /api/v1/auth/change-password:
    post:
      summary: Rotate Credentials With Old Password Proof
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - old_password
                - new_password
              properties:
                old_password:
                  type: string
                new_password:
                  type: string
      responses:
        '200':
          description: Password rotated
        '400':
          description: Invalid request payload
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Invalid old password
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        invalid_params:
          type: object
          additionalProperties:
            type: string