# internal/controlplane/web/handler.go
package web

import (
	"context"
	"fmt"
	"github.com/google/uuid"
	"github.com/ivanchik-byte/Simple-VPN-Builder/internal/controlplane/ai"
	"github.com/ivanchik-byte/Simple-VPN-Builder/internal/controlplane/alerting"
	apimiddleware "github.com/ivanchik-byte/Simple-VPN-Builder/internal/controlplane/api/middleware"
	"github.com/ivanchik-byte/Simple-VPN-Builder/internal/controlplane/api/response"
	"github.com/ivanchik-byte/Simple-VPN-Builder/internal/controlplane/auth"
	cpgrpc "github.com/ivanchik-byte/Simple-VPN-Builder/internal/controlplane/grpc"
	"github.com/ivanchik-byte/Simple-VPN-Builder/internal/controlplane/service"
	"github.com/ivanchik-byte/Simple-VPN-Builder/internal/controlplane/store"
	"github.com/jackc/pgx/v5/pgtype"
	"github.com/skip2/go-qrcode"
	"net"
	"net/http"
	"net/netip"
	"net/url"
	"strings"
	"time"
)

type TelemetryData struct {
	CPUPercent      float64
	CPUModel        string
	RAMPercent      float64
	RAMUsed         int64
	RAMTotal        int64
	DiskPercent     float64
	DiskUsed        int64
	DiskTotal       int64
	RxSpeed         int64
	TxSpeed         int64
	TotalTraffic24h int64
	History         []MetricSnapshot
}

type StatsSummary struct {
	ActiveNodes        int
	TotalNodes         int
	ActiveUsers        int
	TotalUsers         int
	TotalTrafficBytes  int64
	SupportedProtocols int
}

type Handler struct {
	tmpl             *TemplateEngine
	repos            *store.Repositories
	jwtManager       *auth.JWTManager
	passwordManager  *auth.PasswordManager
	totpManager      *auth.TOTPManager
	apiKeyManager    *auth.APIKeyManager
	sessionMgr       *cpgrpc.SessionManager
	provisioner      *service.CredentialProvisioner
	broadcastService *service.BroadcastService
	alertDispatcher  *alerting.AlertDispatcher
	loginLimiter     *apimiddleware.RateLimiter
	copilotService   *ai.CopilotService
}

func (h *Handler) SetCopilotService(c *ai.CopilotService) {
	h.copilotService = c
}

func (h *Handler) SetLoginRateLimiter(rl *apimiddleware.RateLimiter) {
	h.loginLimiter = rl
}

func (h *Handler) SetProvisioner(p *service.CredentialProvisioner) {
	h.provisioner = p
}

func (h *Handler) SetBroadcastService(s *service.BroadcastService) {
	h.broadcastService = s
}

func (h *Handler) SetAlertDispatcher(d *alerting.AlertDispatcher) {
	h.alertDispatcher = d
}

func NewHandler(
	tmpl *TemplateEngine,
	repos *store.Repositories,
	jwtManager *auth.JWTManager,
	passwordManager *auth.PasswordManager,
	totpManager *auth.TOTPManager,
	apiKeyManager *auth.APIKeyManager,
	sessionMgr *cpgrpc.SessionManager,
) *Handler {
	return &Handler{
		tmpl:            tmpl,
		repos:           repos,
		jwtManager:      jwtManager,
		passwordManager: passwordManager,
		totpManager:     totpManager,
		apiKeyManager:   apiKeyManager,
		sessionMgr:      sessionMgr,
	}
}

func (h *Handler) getCallerPermissions(ctx context.Context) store.AdminPermissions {
	adminCtx := GetAdminContext(ctx)
	if adminCtx == nil {
		return store.DefaultAdminPermissions("admin")
	}
	if adminCtx.Role == "owner" {
		return store.DefaultAdminPermissions("owner")
	}
	if h.repos == nil || h.repos.Admins == nil {
		return store.DefaultAdminPermissions(adminCtx.Role)
	}
	callerAdmin, err := h.repos.Admins.GetByID(ctx, adminCtx.AdminID)
	if err != nil {
		return store.DefaultAdminPermissions(adminCtx.Role)
	}
	return callerAdmin.ParsedPermissions()
}

// GET /admin/login

func (h *Handler) Login(w http.ResponseWriter, r *http.Request) {
	if err := r.ParseForm(); err != nil {
		http.Redirect(w, r, "/admin/login?error=Invalid+form+data", http.StatusSeeOther)
		return
	}

	loginInput := strings.TrimSpace(r.FormValue("username"))
	password := r.FormValue("password")
	totpCode := strings.TrimSpace(r.FormValue("totp_code"))

	ctx := r.Context()

	// Extract remote IP for rate limiting
	remoteIP := r.RemoteAddr
	if host, _, err := net.SplitHostPort(r.RemoteAddr); err == nil {
		remoteIP = host
	}

	// 1. Dual-tier brute-force rate limit: Per-IP & Per-Account
	if h.loginLimiter != nil {
		ipKey := "login:ip:" + remoteIP
		if allowed, _, retrySec, _ := h.loginLimiter.Allow(ctx, ipKey); !allowed {
			h.recordAudit(r, "LoginRateLimited", "auth", nil, fmt.Sprintf("IP %s temporarily blocked due to excessive login attempts", remoteIP))
			http.Redirect(w, r, fmt.Sprintf("/admin/login?error=Too+many+login+attempts.+Please+wait+%d+seconds&username=%s", retrySec, url.QueryEscape(loginInput)), http.StatusSeeOther)
			return
		}
		if loginInput != "" {
			userKey := "login:user:" + strings.ToLower(loginInput)
			if allowed, _, retrySec, _ := h.loginLimiter.Allow(ctx, userKey); !allowed {
				h.recordAudit(r, "LoginAccountRateLimited", "auth", nil, fmt.Sprintf("Account %s temporarily blocked due to repeated failures", loginInput))
				http.Redirect(w, r, fmt.Sprintf("/admin/login?error=Account+temporarily+rate-limited.+Please+wait+%d+seconds&username=%s", retrySec, url.QueryEscape(loginInput)), http.StatusSeeOther)
				return
			}
		}
	}

	admin, err := h.repos.Admins.GetByEmail(ctx, loginInput)
	if err != nil && !strings.Contains(loginInput, "@") {
		// Fallback: allow logging in with short username 'admin' matching 'admin@vpnbuilder.local'
		admin, err = h.repos.Admins.GetByEmail(ctx, loginInput+"@vpnbuilder.local")
	}
	if err != nil {
		// Consume constant CPU cycles to prevent timing attacks / user enumeration
		h.passwordManager.DummyVerify(password)
		h.recordAudit(r, "LoginFailed", "auth", nil, fmt.Sprintf("Failed login attempt for unknown user: %s from IP %s", loginInput, remoteIP))
		http.Redirect(w, r, "/admin/login?error=Invalid+credentials&username="+url.QueryEscape(loginInput), http.StatusSeeOther)
		return
	}

	if err := h.passwordManager.Verify(password, admin.PasswordHash); err != nil {
		h.recordAudit(r, "LoginFailed", "auth", &admin.ID, fmt.Sprintf("Invalid password attempt for %s from IP %s", admin.Email, remoteIP))
		http.Redirect(w, r, "/admin/login?error=Invalid+credentials&username="+url.QueryEscape(loginInput), http.StatusSeeOther)
		return
	}

	if admin.TotpSecret.Valid && admin.TotpSecret.String != "" {
		if totpCode == "" || !h.totpManager.ValidateCode(totpCode, admin.TotpSecret.String) {
			h.recordAudit(r, "LoginFailed2FA", "auth", &admin.ID, fmt.Sprintf("Invalid 2FA code attempt for %s from IP %s", admin.Email, remoteIP))
			http.Redirect(w, r, "/admin/login?error=Invalid+2FA+code&username="+url.QueryEscape(loginInput), http.StatusSeeOther)
			return
		}
	}

	roleStr := "admin"
	if admin.Role.Valid {
		roleStr = admin.Role.String
	}

	// Web UI admin session uses 24-hour lifetime to match cookie expiration.
	// The forced-rotation flag travels inside the token so RequireWebAuth can
	// confine must-change sessions to the password-change flow.
	accessToken, err := h.jwtManager.GenerateAccessTokenForAdmin(admin.ID, admin.Email, roleStr, admin.MustChangePassword, 24*time.Hour)
	if err != nil {
		http.Redirect(w, r, "/admin/login?error=Token+generation+failed", http.StatusSeeOther)
		return
	}

	http.SetCookie(w, &http.Cookie{
		Name:     CookieAuthName,
		Value:    accessToken,
		Path:     "/",
		Expires:  time.Now().Add(24 * time.Hour),
		HttpOnly: true,
		Secure:   IsHTTPS(r),
		SameSite: http.SameSiteLaxMode,
	})

	if admin.MustChangePassword {
		h.recordAudit(r, "LoginMustChangePassword", "auth", &admin.ID, fmt.Sprintf("Login with default credentials from IP %s", remoteIP))
		http.Redirect(w, r, "/admin/settings-v2?must_change=1", http.StatusSeeOther)
		return
	}

	http.Redirect(w, r, "/admin/dashboard", http.StatusSeeOther)
}

// POST /admin/logout

func (h *Handler) Logout(w http.ResponseWriter, r *http.Request) {
	cookie, err := r.Cookie(CookieAuthName)
	if err == nil && cookie.Value != "" {
		claims, err := h.jwtManager.ValidateAccessToken(cookie.Value)
		if err == nil && claims != nil && claims.ID != "" && h.jwtManager.Blacklist() != nil {
			_ = h.jwtManager.Blacklist().Revoke(r.Context(), claims.ID, 24*time.Hour)
		}
	}

	http.SetCookie(w, &http.Cookie{
		Name:     CookieAuthName,
		Value:    "",
		Path:     "/",
		MaxAge:   -1,
		HttpOnly: true,
		Secure:   IsHTTPS(r),
		SameSite: http.SameSiteLaxMode,
	})

	http.Redirect(w, r, "/admin/login", http.StatusSeeOther)
}

// GET /admin/dashboard

func (h *Handler) GenerateQR(w http.ResponseWriter, r *http.Request) {
	text := r.URL.Query().Get("text")
	if text == "" {
		http.Error(w, "missing text parameter", http.StatusBadRequest)
		return
	}

	png, err := qrcode.Encode(text, qrcode.Medium, 256)
	if err != nil {
		http.Error(w, "failed to generate QR code", http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "image/png")
	w.WriteHeader(http.StatusOK)
	_, _ = w.Write(png)
}

// GET /client/{token} (deprecated & disabled to hide admin control plane from clients)

func (h *Handler) recordAudit(r *http.Request, action string, resType string, resID *uuid.UUID, details string) {
	if h.repos == nil || h.repos.AuditLogs == nil {
		return
	}

	// Check configurable audit logging category toggles
	if h.repos.Billing != nil {
		if replies, err := h.repos.Billing.GetBotReplies(r.Context()); err == nil {
			cfg := store.ParseLogRetentionSettings(replies)
			if (action == "DirectMessageUser" || action == "SendBroadcast") && !cfg.LogDirectMessages {
				return
			}
			if resType == "auth" && !cfg.LogAuth {
				return
			}
			if resType == "user" && !cfg.LogUserManagement && action != "DirectMessageUser" {
				return
			}
			if resType == "billing" && !cfg.LogBilling {
				return
			}
			if resType == "node" && !cfg.LogNodes {
				return
			}
			if (resType == "settings" || resType == "bot") && !cfg.LogSettings {
				return
			}
		}
	}

	adminCtx := GetAdminContext(r.Context())
	var adminUUID pgtype.UUID
	actorName := "System"
	if adminCtx != nil {
		adminUUID = pgtype.UUID{Bytes: adminCtx.AdminID, Valid: true}
		actorName = adminCtx.Username
	}

	var resourceUUID pgtype.UUID
	targetName := resType
	if resID != nil {
		resourceUUID = pgtype.UUID{Bytes: *resID, Valid: true}
		targetName = fmt.Sprintf("%s:%s", resType, resID.String()[:8])
	}

	ipStr, _, err := net.SplitHostPort(r.RemoteAddr)
	if err != nil {
		ipStr = r.RemoteAddr
	}
	var parsedIP *netip.Addr
	if addr, parseErr := netip.ParseAddr(ipStr); parseErr == nil {
		parsedIP = &addr
	}

	_, _ = h.repos.AuditLogs.Create(r.Context(), store.CreateAuditLogParams{
		AdminID:      adminUUID,
		ApiKeyID:     pgtype.UUID{Valid: false},
		Action:       action,
		ResourceType: pgtype.Text{String: resType, Valid: resType != ""},
		ResourceID:   resourceUUID,
		Diff:         []byte(details),
		IpAddress:    parsedIP,
		UserAgent:    pgtype.Text{String: r.UserAgent(), Valid: r.UserAgent() != ""},
	})

	if h.alertDispatcher != nil {
		h.alertDispatcher.SendAuditAlert(actorName, action, targetName, details)
	}
}

// POST /admin/users/{id}/ban

func (h *Handler) NotFound(w http.ResponseWriter, r *http.Request) {
	if strings.HasPrefix(r.URL.Path, "/api/") {
		response.RespondNotFound(w, r, "Resource not found")
		return
	}
	w.WriteHeader(http.StatusNotFound)
	_ = h.tmpl.RenderStandalone(w, "error.html", map[string]any{
		"Code":    "404",
		"Title":   "Page Not Found",
		"Message": "The requested endpoint or resource was not found on this server.",
		"Accent":  "#a78bfa",
		"Glow":    "rgba(139,92,246,0.12)",
	})
}

// RenderErrorPage writes a standalone error document with the given status.

func (h *Handler) RenderErrorPage(w http.ResponseWriter, code int, title, message, accent, glow string, retryAfter int) {
	w.Header().Set("Content-Type", "text/html; charset=utf-8")
	w.WriteHeader(code)
	_ = h.tmpl.RenderStandalone(w, "error.html", map[string]any{
		"Code":       code,
		"Title":      title,
		"Message":    message,
		"Accent":     accent,
		"Glow":       glow,
		"RetryAfter": retryAfter,
	})
}

// ChangePassword rotates the current admin password (cookie session required).
func (h *Handler) ChangePassword(w http.ResponseWriter, r *http.Request) {
	adminCtx := GetAdminContext(r.Context())
	if adminCtx == nil {
		http.Redirect(w, r, "/admin/login", http.StatusSeeOther)
		return
	}

	if err := r.ParseForm(); err != nil {
		http.Redirect(w, r, "/admin/settings-v2?error=invalid_form", http.StatusSeeOther)
		return
	}

	oldPass := r.FormValue("old_password")
	newPass := r.FormValue("new_password")
	if oldPass == "" || len(newPass) < 12 {
		http.Redirect(w, r, "/admin/settings-v2?error=Password+must+be+12+characters+or+longer", http.StatusSeeOther)
		return
	}

	admin, err := h.repos.Admins.GetByID(r.Context(), adminCtx.AdminID)
	if err != nil {
		http.Redirect(w, r, "/admin/login", http.StatusSeeOther)
		return
	}

	if err := h.passwordManager.Verify(oldPass, admin.PasswordHash); err != nil {
		h.recordAudit(r, "ChangePasswordFailed", "auth", &admin.ID, "Wrong current password")
		http.Redirect(w, r, "/admin/settings-v2?error=Current+password+is+incorrect", http.StatusSeeOther)
		return
	}

	hash, err := h.passwordManager.Hash(newPass)
	if err != nil {
		http.Redirect(w, r, "/admin/settings-v2?error=Hashing+failed", http.StatusSeeOther)
		return
	}

	if err := h.repos.Admins.UpdatePassword(r.Context(), admin.ID, hash); err != nil {
		http.Redirect(w, r, "/admin/settings-v2?error=Password+update+failed", http.StatusSeeOther)
		return
	}

	// Rotation hygiene: invalidate every token ever issued for this admin and
	// drop every API key they created. The current cookie session dies here,
	// so the caller must log in again with the new password.
	if h.jwtManager != nil && h.jwtManager.Blacklist() != nil {
		_ = h.jwtManager.Blacklist().RevokeAdmin(r.Context(), admin.ID.String(), 7*24*time.Hour)
	}
	if h.repos != nil && h.repos.APIKeys != nil {
		_ = h.repos.APIKeys.DeleteByAdminID(r.Context(), admin.ID)
	}

	h.recordAudit(r, "ChangePassword", "auth", &admin.ID, "Password rotated")
	http.Redirect(w, r, "/admin/login?message=Password+changed.+Please+log+in+again", http.StatusSeeOther)
}